// src/api/openapi/models/SeedlingRequestListItem.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type SeedlingRequestListItem = {
    id?: string;
    crop_id?: string;
    farm_id?: string;
    other?: string;
    quantity_request?: string;
    quantity_approve?: string;
    status?: string;
    delivery_date?: string;
    note?: string;
    name?: string;
    farm_name?: string;
};